refactor(mcp): converge OAuth integration ownership
Move concrete RMCP authorization and credential access behind the services integration owner. Keep Core's compatibility facade and product callback orchestration while routing Desktop credential status through the assembled manager.
